requirements.txt
requests
tqdm
pymupdf
pymongo==3.12
numpy
sentence-transformers
rank-bm25
fastapi
uvicorn
python-dotenv
